backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
asgiref==3.9.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   django-cors-headers
attrs==25.3.0
    # via
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==25.4.16
    # via twisted
beautifulsoup4==4.13.4
    # via -r requirements.in
boto3==1.39.13
    # via -r requirements.in
botocore==1.39.13
    # via
    #   boto3
    #   s3transfer
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
channels[daphne]==4.2.2
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.3.0
    # via -r requirements.in
charset-normalizer==3.4.2
    # via requests
constantly==23.10.4
    # via twisted
cryptography==45.0.5
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.2.1
    # via channels
distro==1.9.0
    # via openai
dj-database-url==3.0.1
    # via -r requirements.in
django==4.2.2
    # via
    #   -r requirements.in
    #   channels
    #   dj-database-url
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
django-cors-headers==4.7.0
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
django-ratelimit==4.1.0
    # via -r requirements.in
djangorestframework==3.16.0
    # via -r requirements.in
google-api-core==2.25.1
    # via google-api-python-client
google-api-python-client==2.177.0
    # via -r requirements.in
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.2
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via google-api-core
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via openai
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.10
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
    #   twisted
incremental==24.7.2
    # via twisted
jiter==0.10.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markdown==3.8.2
    # via -r requirements.in
msgpack==1.1.1
    # via channels-redis
oauthlib==3.3.1
    # via requests-oauthlib
openai==1.97.1
    # via -r requirements.in
packaging==25.0
    # via gunicorn
proto-plus==1.26.1
    # via google-api-core
protobuf==6.31.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg[binary,pool]==3.2.9
    # via -r requirements.in
psycopg-binary==3.2.9
    # via psycopg
psycopg-pool==3.2.6
    # via psycopg
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
    #   service-identity
pyasn1-modules==0.4.2
    # via
    #   google-auth
    #   service-identity
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via openai
pydantic-core==2.33.2
    # via pydantic
pyjwt==2.10.1
    # via -r requirements.in
pyopenssl==25.1.0
    # via twisted
pyparsing==3.2.3
    # via httplib2
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.1.1
    # via -r requirements.in
redis==6.2.0
    # via
    #   -r requirements.in
    #   channels-redis
requests==2.32.4
    # via
    #   google-api-core
    #   requests-oauthlib
    #   stripe
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
s3transfer==0.13.1
    # via boto3
service-identity==24.2.0
    # via twisted
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.7
    # via beautifulsoup4
sqlparse==0.5.3
    # via django
stripe==12.3.0
    # via -r requirements.in
tqdm==4.67.1
    # via openai
twisted[tls]==25.5.0
    # via daphne
txaio==25.6.1
    # via autobahn
typing-extensions==4.14.1
    # via
    #   beautifulsoup4
    #   openai
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
    #   stripe
    #   twisted
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via
    #   botocore
    #   requests
whitenoise==6.9.0
    # via -r requirements.in
zope-interface==7.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
